requirements.txt
streamlit==1.46.1
streamlit-webrtc==0.63.3
tensorflow==2.15.0
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
pillow
